refactor: colocate v3 operators with their layout functions
Move v3 operator definitions (spread/stack/scatter/table/group) from chart.ts
into the same file as their low-level layout, mirroring how createMark is
called at the bottom of each shape file (e.g. rect.tsx: `Rect` + `rect`).

The factory's signature is now: createOperator(layout, { split, channels }).
- split: (opts, data) -> Map<key, Datum[]>  (or {entries, keys} for table)
- channels: { w: "size" } | { x: { type: "pos", entry: true } }
- enumerateMarks gone; combinator form's marks: Mark<T>[] always, factory
  enumerates with integer keys.
- Meta gone; SplitResult either a bare Map or {entries, keys} for axis
  labels.
- DeriveOperatorOptions utility type exported (mirrors DeriveMarkProps);
  scatter retains a single documented cast pending full wiring.

Naming convention now mirrors shapes:
- packages/gofish-graphics/src/ast/graphicalOperators/spread.tsx exports
  `Spread` (low-level) and `spread`/`stack` (v3) plus SpreadOptions.
- scatter.tsx: `Scatter` + `scatter` + ScatterOptions.
- table.tsx: `Table` + `table` + TableOptions.
- frame.tsx: `Frame` + `group` + GroupOptions.

ChartBuilder extracted to packages/gofish-graphics/src/ast/marks/chartBuilder.ts
to keep createOperator's import graph free of chart.ts dependencies. The
runtime import cycle (chartBuilder <-> spread <-> createOperator) resolves
cleanly via ES-module live bindings — references are only used inside method
bodies.

Internal callers (spreadX/spreadY/stack.tsx in graphicalOperators/) updated
to import the now-capitalized `Spread`. lib.ts re-exports both forms from
each operator file.

Build passes. ~230 v3 call sites in stories/tests/docs examples and the
Python wrapper still need migration to the new signatures (next step).
