+++ packages/gofish-graphics/src/ast/graphicalOperators/frame.tsx
@@ -4,7 +4,7 @@ import { coord } from "../coordinateTransforms/coord";
 import { layer } from "./layer";
 import { createNodeOperator } from "../withGoFish";
 import { GoFishAST } from "../_ast";
-export const frame = createNodeOperator(
+export const Frame = createNodeOperator(
   (
     options: {
       key?: string;
@@ -29,5 +29,22 @@ export const frame = createNodeOperator(
     } else {
       return layer(options, children);
     }
+  }
+);
+
+import { createOperator } from "../marks/createOperator";
+
+export type GroupOptions = {
+  by?: string;
+};
+
+export const group = createOperator<any, GroupOptions>(
+  async (_opts, children) =>
+    (await Frame({}, children)) as unknown as GoFishAST,
+  {
+    split: ({ by }, d) => {
+      if (!by) throw new Error("group requires opts.by = fieldName");
+      return Map.groupBy(d, (r: any) => r[by]);
+    },
   }
 );
