+++ packages/gofish-graphics/src/ast/graphicalOperators/spread.tsx
@@ -34,7 +34,7 @@ const unwrapLodashArray = function <T>(value: T[] | Collection<T>): T[] {
   return value as T[];
 };
 
-export const spread = createNodeOperator(
+export const Spread = createNodeOperator(
   (
     {
       name,
@@ -410,3 +410,45 @@ export const spread = createNodeOperator(
     );
   }
 );
+
+import { createOperator, LayoutFn } from "../marks/createOperator";
+import { Mark, Operator } from "../types";
+
+export type SpreadOptions<T = any> = {
+  by?: keyof T & string;
+  dir: "x" | "y";
+  spacing?: number;
+  alignment?: "start" | "middle" | "end" | "baseline";
+  sharedScale?: boolean;
+  mode?: "edge" | "center";
+  reverse?: boolean;
+  w?: number | (keyof T & string);
+  h?: number | (keyof T & string);
+  debug?: boolean;
+};
+
+export const spread = createOperator<any, SpreadOptions>(
+  Spread as unknown as LayoutFn<SpreadOptions>,
+  {
+    split: ({ by }, d) =>
+      by
+        ? Map.groupBy(d, (r: any) => r[by])
+        : new Map(d.map((r, i) => [i, [r]])),
+    channels: { w: "size", h: "size" },
+  }
+);
+
+/** Stack has no `spacing` option — children always touch (spacing: 0). */
+export type StackOptions<T = any> = Omit<SpreadOptions<T>, "spacing">;
+
+export function stack(
+  opts: StackOptions,
+  marks: Mark<any>[]
+): ReturnType<typeof spread>;
+export function stack(opts: StackOptions): Operator<any[], any[]>;
+export function stack(opts: StackOptions, marks?: Mark<any>[]): any {
+  const stackOpts = { ...opts, spacing: 0 } as SpreadOptions;
+  return marks !== undefined
+    ? (spread as any)(stackOpts, marks)
+    : (spread as any)(stackOpts);
+}
