+++ packages/gofish-graphics/src/ast/graphicalOperators/stack.tsx
@@ -1,4 +1,4 @@
-import { spread } from "./spread";
+import { Spread } from "./spread";
 import { FancyDims } from "../dims";
 import { MaybeValue } from "../data";
 
@@ -15,8 +15,8 @@ type StackProps = {
 export const stack = (
   props: StackProps,
   children: any // GoFishChildrenInput type from ../withGoFish is not exported, using any
-): ReturnType<typeof spread> => {
-  return spread(
+): ReturnType<typeof Spread> => {
+  return Spread(
     {
       ...props,
       spacing: 0,
