+++ packages/gofish-graphics/src/lib.ts
@@ -49,8 +49,8 @@ export type {
 export { stack as Stack } from "./ast/graphicalOperators/stack";
 export { stackX as StackX } from "./ast/graphicalOperators/stackX";
 export { stackY as StackY } from "./ast/graphicalOperators/stackY";
-export { spread as Spread } from "./ast/graphicalOperators/spread";
-export { scatter as Scatter } from "./ast/graphicalOperators/scatter";
+export { Spread, spread, stack } from "./ast/graphicalOperators/spread";
+export { Scatter, scatter } from "./ast/graphicalOperators/scatter";
 export { spreadX as SpreadX } from "./ast/graphicalOperators/spreadX";
 export { spreadY as SpreadY } from "./ast/graphicalOperators/spreadY";
 export { layer as Layer } from "./ast/graphicalOperators/layer";
@@ -59,10 +59,10 @@ export { connect as Connect } from "./ast/graphicalOperators/connect";
 export { connectX as ConnectX } from "./ast/graphicalOperators/connectX";
 export { connectY as ConnectY } from "./ast/graphicalOperators/connectY";
 export { enclose as Enclose } from "./ast/graphicalOperators/enclose";
-export { frame as Frame } from "./ast/graphicalOperators/frame";
+export { Frame, group } from "./ast/graphicalOperators/frame";
 export { position as Position } from "./ast/graphicalOperators/position";
 export { arrow as Arrow } from "./ast/graphicalOperators/arrow";
-export { table as Table } from "./ast/graphicalOperators/table";
+export { Table, table } from "./ast/graphicalOperators/table";
 export {
   over as Over,
   inside as In,
@@ -81,11 +81,6 @@ export { image } from "./ast/shapes/image";
 /* Chart Syntax */
 export {
   chart as Chart,
-  spread,
-  stack,
-  table,
-  scatter,
-  group,
   derive,
   rect,
   circle,
